refactor(mutations): CardType handlers own their cascades
Move the CardType cascades out of CardTypeResource and into their owning
mutation handlers, mirroring LinkTypeRenameHandler and the just-completed
leaf-resource family. Resource-level ops now do only validation, content
building and pure file I/O; the engine handlers own all cross-resource and
card-metadata cascades.

card-type-rename.ts (CardTypeRenameHandler):
- already rewrote each card's metadata.cardType; now also performs the
  calculation / report-handlebar / card-content reference rewrites (via
  cascades/rewrite-refs.ts) and rewrites every local link type's
  sourceCardTypes / destinationCardTypes (new private updateLinkTypes),
  all BEFORE resource.rename().

card-type-add-custom-field.ts (CardTypeAddCustomFieldHandler):
- after resource.update validates + persists the field, writes the new
  field as null on every affected card (project + local template cards).

card-type-remove-custom-field.ts (CardTypeRemoveCustomFieldHandler):
- after resource.update removes the field (and strips it from the
  visible-fields arrays), drops the field key from every affected card.

card-type-workflow-change.ts (CardTypeWorkflowChangeHandler):
- now owns verifyStateMapping (runs BEFORE the resource write so an invalid
  mapping rejects without persisting) and the per-card workflowState remap.

card-type-resource.ts:
- update() for workflow/customFields no longer triggers a cascade; it only
  builds content, validates and writes via postUpdate.
- onNameChange() keeps only the self-prefix rewrites of its own
  customFields / alwaysVisibleFields / optionallyVisibleFields / workflow
  plus write(); the cross-resource cascade is gone.
- deleted handleCustomFieldsChange, handleAddNewField, handleRemoveField,
  handleWorkflowChange, verifyStateMapping, updateLinkTypes and the private
  updateCardMetadata helper, plus now-unused imports.
- CardType already declared a concrete onNameChange, so no new no-op stub
  was needed (FolderResource/FileResource base-cleanup is a later step).

resources.test.ts: the resource-level "should update card contents when
renaming card type" case asserted the resource rename cascaded card content.
It now asserts the resource class no longer rewrites references (renamed to
"...via resource class (cascade moved to handler)"), matching the link-type
and leaf cases. The positive cascade coverage lives in the handler tests
(test/mutations/handlers/card-type-*.test.ts), which exercise the cascade
through ResourceMutations and remain green.
